pyproject.toml
[project]
name = "google-cloud-bucket-viewer"
version = "0.1.0"
description = "A web application to view images from Google Cloud Storage buckets"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pytest>=8.4.2",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "google-cloud-storage>=2.18.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
]
